Update README and other docs. Rename AdaptiveComparison -> AdaptiveBetting Use OneAudit max99 for compare audit type plots.
