+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/workflow/TestRaireWorkflow.kt
@@ -18,7 +18,7 @@ class TestRaireWorkflow {
     }
 
     fun testRaireWorkflow(auditConfig: AuditConfig) {
-        val (rcontest, testCvrs) = makeRaireContest(N=20000, contestId=111, ncands=3, minMargin=.04, quiet = true)
+        val (rcontest, testCvrs) = simulateRaireTestData(N=20000, contestId=111, ncands=3, minMargin=.04, quiet = true)
         val workflow = ClcaWorkflow(auditConfig, emptyList(), listOf(rcontest), BallotCardsClcaStart(testCvrs, testCvrs, auditConfig.seed))
         runWorkflow("testRaireWorkflow", workflow)
     }
@@ -29,7 +29,7 @@ class TestRaireWorkflow {
         val auditConfig = AuditConfig(AuditType.CLCA, hasStyles=false, nsimEst=10,
             clcaConfig = ClcaConfig(ClcaStrategyType.fuzzPct, simFuzzPct = mvrFuzzPct))
 
-        val (rcontest: RaireContestUnderAudit, testCvrs: List<Cvr>) = makeRaireContest(N=20000, contestId=111, ncands=4, minMargin=.04, quiet = true)
+        val (rcontest: RaireContestUnderAudit, testCvrs: List<Cvr>) = simulateRaireTestData(N=20000, contestId=111, ncands=4, minMargin=.04, quiet = true)
         val testMvrs = makeFuzzedCvrsFrom(listOf(rcontest.contest), testCvrs, mvrFuzzPct)
         val workflow = ClcaWorkflow(auditConfig, emptyList(), listOf(rcontest), BallotCardsClcaStart(testCvrs, testMvrs, auditConfig.seed))
         runWorkflow("testRaireWorkflow", workflow)
