+++ docs/plots/samples/auditsWithErrors/auditsWithErrorsLogLog.html
@@ -5,7 +5,7 @@
        &lt;script type=&quot;text/javascript&quot; data-lets-plot-script=&quot;library&quot; src=&quot;https://cdn.jsdelivr.net/gh/JetBrains/lets-plot@v4.5.1/js-package/distr/lets-plot.min.js&quot;>&lt;/script>
    &lt;/head>
    &lt;body>
-          &lt;div id=&quot;rcnIGr&quot;>&lt;/div>
+          &lt;div id=&quot;PloAuh&quot;>&lt;/div>
    &lt;script type=&quot;text/javascript&quot; data-lets-plot-script=&quot;plot&quot;>
    
    (function() {
@@ -23,7 +23,8 @@
 &quot;kind&quot;:&quot;plot&quot;,
 &quot;scales&quot;:[{
 &quot;aesthetic&quot;:&quot;x&quot;,
-&quot;limits&quot;:[null,null]
+&quot;limits&quot;:[null,null],
+&quot;trans&quot;:&quot;LOG10&quot;
 },{
 &quot;aesthetic&quot;:&quot;y&quot;,
 &quot;limits&quot;:[null,null],
@@ -33,7 +34,8 @@
 &quot;discrete&quot;:true
 },{
 &quot;aesthetic&quot;:&quot;x&quot;,
-&quot;limits&quot;:[null,null]
+&quot;limits&quot;:[null,null],
+&quot;trans&quot;:&quot;LOG10&quot;
 },{
 &quot;aesthetic&quot;:&quot;y&quot;,
 &quot;limits&quot;:[null,null],
@@ -112,9 +114,9 @@
 &quot;disable_splitting&quot;:true
 }
 }],
-&quot;spec_id&quot;:&quot;6&quot;
+&quot;spec_id&quot;:&quot;9&quot;
 };
-       var containerDiv = document.getElementById(&quot;rcnIGr&quot;);
+       var containerDiv = document.getElementById(&quot;PloAuh&quot;);
        
                var toolbar = null;
                var plotContainer = containerDiv;
