+++ docs/plots/samples/auditsWithPhantoms/auditsWithPhantomsLogLinear.html
@@ -5,7 +5,7 @@
        &lt;script type=&quot;text/javascript&quot; data-lets-plot-script=&quot;library&quot; src=&quot;https://cdn.jsdelivr.net/gh/JetBrains/lets-plot@v4.5.1/js-package/distr/lets-plot.min.js&quot;>&lt;/script>
    &lt;/head>
    &lt;body>
-          &lt;div id=&quot;HsP9uT&quot;>&lt;/div>
+          &lt;div id=&quot;9Zs4Z5&quot;>&lt;/div>
    &lt;script type=&quot;text/javascript&quot; data-lets-plot-script=&quot;plot&quot;>
    
    (function() {
@@ -48,10 +48,10 @@
 },
 &quot;stat&quot;:&quot;identity&quot;,
 &quot;data&quot;:{
-&quot;&amp;merged_groups&quot;:[&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;],
-&quot;samplesNeeded&quot;:[222.05,365.0099999999999,587.1999999999999,743.6999999999998,3390.1000000000004,14726.110000000004,50000.0,6137.779999999999,6991.960000000001,8829.899999999996,13405.249999999998,19913.75000000001,50000.0,50000.0,3837.9600000000005,4748.339999999999,5937.409999999998,9451.350000000002,22737.800000000003,50000.0,50000.0],
-&quot;phantomPct&quot;:[0.0,0.005,0.01,0.02,0.03,0.04,0.05,0.0,0.005,0.01,0.02,0.03,0.04,0.05,0.0,0.005,0.01,0.02,0.03,0.04,0.05],
-&quot;auditType&quot;:[&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;]
+&quot;&amp;merged_groups&quot;:[&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;],
+&quot;samplesNeeded&quot;:[295.3599999999999,350.1200000000001,400.31,528.8000000000001,1818.2700000000002,13015.209999999997,50000.0,261.35999999999984,372.92999999999995,639.49,4133.580000000001,29009.170000000002,45435.63,50000.0,3738.559999999999,4777.070000000001,5457.199999999998,9059.710000000006,18533.510000000002,41558.23000000002,46593.92],
+&quot;auditType&quot;:[&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;],
+&quot;phantomPct&quot;:[0.0,0.005,0.01,0.02,0.03,0.04,0.05,0.0,0.005,0.01,0.02,0.03,0.04,0.05,0.0,0.005,0.01,0.02,0.03,0.04,0.05]
 },
 &quot;sampling&quot;:&quot;none&quot;,
 &quot;inherit_aes&quot;:false,
@@ -80,10 +80,10 @@
 },
 &quot;stat&quot;:&quot;identity&quot;,
 &quot;data&quot;:{
-&quot;&amp;merged_groups&quot;:[&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;],
-&quot;samplesNeeded&quot;:[222.05,365.0099999999999,587.1999999999999,743.6999999999998,3390.1000000000004,14726.110000000004,50000.0,6137.779999999999,6991.960000000001,8829.899999999996,13405.249999999998,19913.75000000001,50000.0,50000.0,3837.9600000000005,4748.339999999999,5937.409999999998,9451.350000000002,22737.800000000003,50000.0,50000.0],
-&quot;phantomPct&quot;:[0.0,0.005,0.01,0.02,0.03,0.04,0.05,0.0,0.005,0.01,0.02,0.03,0.04,0.05,0.0,0.005,0.01,0.02,0.03,0.04,0.05],
-&quot;auditType&quot;:[&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;oneaudit&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;]
+&quot;&amp;merged_groups&quot;:[&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;],
+&quot;samplesNeeded&quot;:[295.3599999999999,350.1200000000001,400.31,528.8000000000001,1818.2700000000002,13015.209999999997,50000.0,261.35999999999984,372.92999999999995,639.49,4133.580000000001,29009.170000000002,45435.63,50000.0,3738.559999999999,4777.070000000001,5457.199999999998,9059.710000000006,18533.510000000002,41558.23000000002,46593.92],
+&quot;auditType&quot;:[&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;clca&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;oneaudit99&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;,&quot;polling&quot;],
+&quot;phantomPct&quot;:[0.0,0.005,0.01,0.02,0.03,0.04,0.05,0.0,0.005,0.01,0.02,0.03,0.04,0.05,0.0,0.005,0.01,0.02,0.03,0.04,0.05]
 },
 &quot;size&quot;:1.0,
 &quot;shape&quot;:1.0,
@@ -114,7 +114,7 @@
 }],
 &quot;spec_id&quot;:&quot;6&quot;
 };
-       var containerDiv = document.getElementById(&quot;HsP9uT&quot;);
+       var containerDiv = document.getElementById(&quot;9Zs4Z5&quot;);
        
                var toolbar = null;
                var plotContainer = containerDiv;
