--- docs/plots/samples/auditsWithUndervotes/auditsWithUndervotes2.cvs
@@ -1,34 +1,0 @@
-parameters, N, margin, status, nrounds, samplesUsed, samplesNeeded, nmvrs, failPct, neededStddev
-"nruns=100 undervote=0.05 cat=polling fuzzPct=0.01 auditType=2.0 ", 50000, 0.02, StatRejectNull, 1.15, 13566.42, 13566.420000000002, 22392.98, 0.0, 6511.561961895165
-"nruns=100 undervote=0.2 cat=polling fuzzPct=0.01 auditType=2.0 ", 50000, 0.02, StatRejectNull, 1.19, 11596.45, 11596.449999999999, 20432.81, 0.0, 6302.438066930924
-"nruns=100 undervote=0.1 cat=polling fuzzPct=0.01 auditType=2.0 ", 50000, 0.02, StatRejectNull, 1.19, 11997.64, 11997.639999999992, 21747.74, 0.0, 6827.311587323375
-"nruns=100 undervote=0.3 cat=polling fuzzPct=0.01 auditType=2.0 ", 50000, 0.02, StatRejectNull, 1.19, 10778.89, 10778.89, 18443.88, 0.0, 6186.475781727429
-"nruns=100 undervote=0.4 cat=polling fuzzPct=0.01 auditType=2.0 ", 50000, 0.02, StatRejectNull, 1.2, 10184.04, 10184.04, 17177.16, 0.0, 5711.460585384444
-"nruns=100 undervote=0.5 cat=polling fuzzPct=0.01 auditType=2.0 ", 50000, 0.02, StatRejectNull, 1.22, 10144.44, 10144.440000000002, 15883.43, 0.0, 5499.645023671983
-"nruns=100 undervote=0.0 cat=polling fuzzPct=0.01 auditType=2.0 ", 50000, 0.02, StatRejectNull, 1.21, 14049.73, 14049.729999999996, 24180.48, 0.0, 7031.395098918849
-"nruns=100 undervote=0.45 cat=polling fuzzPct=0.01 auditType=2.0 ", 50000, 0.02, StatRejectNull, 1.21, 10007.37, 10007.369999999997, 16608.75, 0.0, 5502.9897831179005
-"nruns=100 undervote=0.15 cat=polling fuzzPct=0.01 auditType=2.0 ", 50000, 0.02, StatRejectNull, 1.22, 12248.67, 12248.67, 21669.14, 0.0, 7266.2034434152765
-"nruns=100 undervote=0.25 cat=polling fuzzPct=0.01 auditType=2.0 ", 50000, 0.02, StatRejectNull, 1.22, 12166.47, 12166.47, 19504.1, 0.0, 5499.401004573134
-"nruns=100 undervote=0.35 cat=polling fuzzPct=0.01 auditType=2.0 ", 50000, 0.02, StatRejectNull, 1.25, 10511.16, 10511.159999999998, 18301.83, 0.0, 6353.973293491247
-"nruns=100 undervote=0.1 cat=clca mvrsFuzzPct=0.01 auditType=3.0 ", 50000, 0.02, StatRejectNull, 2.76, 1108.17, 1108.1700000000003, 1267.44, 0.0, 1040.5426474200854
-"nruns=100 undervote=0.5 cat=clca mvrsFuzzPct=0.01 auditType=3.0 ", 50000, 0.02, StatRejectNull, 2.07, 683.52, 683.5199999999999, 762.8, 0.0, 649.283181362339
-"nruns=100 undervote=0.45 cat=clca mvrsFuzzPct=0.01 auditType=3.0 ", 50000, 0.02, StatRejectNull, 2.16, 723.46, 723.4599999999999, 803.59, 0.0, 686.2385215069174
-"nruns=100 undervote=0.1 cat=oneaudit cvrPercent=0.5 fuzzPct=0.01 auditType=1.0 ", 50000, 0.02, StatRejectNull, 1.18, 19629.18, 19629.17999999999, 25399.72, 0.0, 4434.964704211299
-"nruns=100 undervote=0.2 cat=clca mvrsFuzzPct=0.01 auditType=3.0 ", 50000, 0.02, StatRejectNull, 2.26, 772.14, 772.14, 867.67, 0.0, 817.3183225647153
-"nruns=100 undervote=0.3 cat=clca mvrsFuzzPct=0.01 auditType=3.0 ", 50000, 0.02, StatRejectNull, 2.61, 1008.46, 1008.4600000000002, 1130.06, 0.0, 870.9802916254765
-"nruns=100 undervote=0.25 cat=clca mvrsFuzzPct=0.01 auditType=3.0 ", 50000, 0.02, StatRejectNull, 2.27, 805.1, 805.1000000000001, 913.4, 0.0, 788.5087000154149
-"nruns=100 undervote=0.35 cat=clca mvrsFuzzPct=0.01 auditType=3.0 ", 50000, 0.02, StatRejectNull, 2.41, 823.77, 823.7700000000003, 938.84, 0.0, 753.6075484627262
-"nruns=100 undervote=0.4 cat=clca mvrsFuzzPct=0.01 auditType=3.0 ", 50000, 0.02, StatRejectNull, 2.37, 856.27, 856.2699999999999, 951.56, 0.0, 809.6475511603796
-"nruns=100 undervote=0.05 cat=oneaudit cvrPercent=0.5 fuzzPct=0.01 auditType=1.0 ", 50000, 0.02, StatRejectNull, 1.18, 19675.93, 19675.929999999993, 26018.39, 0.0, 4635.656331642802
-"nruns=100 undervote=0.0 cat=clca mvrsFuzzPct=0.01 auditType=3.0 ", 50000, 0.02, StatRejectNull, 2.42, 941.83, 941.8300000000003, 1050.22, 0.0, 891.3594455100589
-"nruns=100 undervote=0.35 cat=oneaudit cvrPercent=0.5 fuzzPct=0.01 auditType=1.0 ", 50000, 0.02, StatRejectNull, 1.24, 20045.55, 20045.550000000007, 24963.66, 0.0, 3800.207074292137
-"nruns=100 undervote=0.15 cat=clca mvrsFuzzPct=0.01 auditType=3.0 ", 50000, 0.02, StatRejectNull, 2.61, 980.59, 980.5900000000003, 1117.74, 0.0, 952.912914121747
-"nruns=100 undervote=0.4 cat=oneaudit cvrPercent=0.5 fuzzPct=0.01 auditType=1.0 ", 50000, 0.02, StatRejectNull, 1.22, 19371.42, 19371.419999999995, 24770.82, 0.0, 4002.802166432911
-"nruns=100 undervote=0.0 cat=oneaudit cvrPercent=0.5 fuzzPct=0.01 auditType=1.0 ", 50000, 0.02, StatRejectNull, 1.18, 19989.59, 19989.589999999993, 26021.8, 0.0, 4657.593921962283
-"nruns=100 undervote=0.3 cat=oneaudit cvrPercent=0.5 fuzzPct=0.01 auditType=1.0 ", 50000, 0.02, StatRejectNull, 1.24, 19998.81, 19998.81, 25746.16, 0.0, 4432.641927101716
-"nruns=100 undervote=0.45 cat=oneaudit cvrPercent=0.5 fuzzPct=0.01 auditType=1.0 ", 50000, 0.02, StatRejectNull, 1.29, 19905.93, 19905.93, 25206.42, 0.0, 3981.857921259872
-"nruns=100 undervote=0.15 cat=oneaudit cvrPercent=0.5 fuzzPct=0.01 auditType=1.0 ", 50000, 0.02, StatRejectNull, 1.21, 19813.37, 19813.36999999999, 26012.47, 0.0, 4491.013156638488
-"nruns=100 undervote=0.25 cat=oneaudit cvrPercent=0.5 fuzzPct=0.01 auditType=1.0 ", 50000, 0.02, StatRejectNull, 1.25, 20122.7, 20122.700000000008, 26126.95, 0.0, 4823.73214119524
-"nruns=100 undervote=0.5 cat=oneaudit cvrPercent=0.5 fuzzPct=0.01 auditType=1.0 ", 50000, 0.02, StatRejectNull, 1.29, 19859.67, 19859.669999999995, 24656.39, 0.0, 3956.181960059472
-"nruns=100 undervote=0.2 cat=oneaudit cvrPercent=0.5 fuzzPct=0.01 auditType=1.0 ", 50000, 0.02, StatRejectNull, 1.33, 20877.35, 20877.349999999995, 27133.44, 0.0, 4962.590632673625
-"nruns=100 undervote=0.05 cat=clca mvrsFuzzPct=0.01 auditType=3.0 ", 50000, 0.02, StatRejectNull, 3.13, 1280.69, 1280.6899999999994, 1440.6, 0.0, 927.7260769753104
