+++ docs/plots/samples/phantomMarginShift/phantomMarginShiftLogLog.html
@@ -5,7 +5,7 @@
        &lt;script type=&quot;text/javascript&quot; data-lets-plot-script=&quot;library&quot; src=&quot;https://cdn.jsdelivr.net/gh/JetBrains/lets-plot@v4.5.1/js-package/distr/lets-plot.min.js&quot;>&lt;/script>
    &lt;/head>
    &lt;body>
-          &lt;div id=&quot;3dq3y9&quot;>&lt;/div>
+          &lt;div id=&quot;0m1sRp&quot;>&lt;/div>
    &lt;script type=&quot;text/javascript&quot; data-lets-plot-script=&quot;plot&quot;>
    
    (function() {
@@ -23,7 +23,8 @@
 &quot;kind&quot;:&quot;plot&quot;,
 &quot;scales&quot;:[{
 &quot;aesthetic&quot;:&quot;x&quot;,
-&quot;limits&quot;:[null,null]
+&quot;limits&quot;:[null,null],
+&quot;trans&quot;:&quot;LOG10&quot;
 },{
 &quot;aesthetic&quot;:&quot;y&quot;,
 &quot;limits&quot;:[null,null],
@@ -33,7 +34,8 @@
 &quot;discrete&quot;:true
 },{
 &quot;aesthetic&quot;:&quot;x&quot;,
-&quot;limits&quot;:[null,null]
+&quot;limits&quot;:[null,null],
+&quot;trans&quot;:&quot;LOG10&quot;
 },{
 &quot;aesthetic&quot;:&quot;y&quot;,
 &quot;limits&quot;:[null,null],
@@ -50,8 +52,8 @@
 &quot;data&quot;:{
 &quot;&amp;merged_groups&quot;:[&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;],
 &quot;samplesNeeded&quot;:[277.1600000000001,371.0099999999999,407.4400000000001,816.4399999999999,1936.0500000000002,4306.999999999999,11917.449999999997,22277.709999999995,319.75,407.7400000000001,750.92,1133.1599999999996,3036.4700000000003,7387.16,25500.78,43940.93],
-&quot;phantomPct&quot;:[0.0,0.005,0.01,0.02,0.03,0.035,0.04,0.0425,0.0,0.005,0.01,0.02,0.03,0.035,0.04,0.0425],
-&quot;auditType&quot;:[&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;]
+&quot;auditType&quot;:[&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;marginShift&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;,&quot;phantoms&quot;],
+&quot;phantomPct&quot;:[0.0,0.005,0.01,0.02,0.03,0.035,0.04,0.0425,0.0,0.005,0.01,0.02,0.03,0.035,0.04,0.0425]
 },
 &quot;sampling&quot;:&quot;none&quot;,
 &quot;inherit_aes&quot;:false,
@@ -112,9 +114,9 @@
 &quot;disable_splitting&quot;:true
 }
 }],
-&quot;spec_id&quot;:&quot;6&quot;
+&quot;spec_id&quot;:&quot;9&quot;
 };
-       var containerDiv = document.getElementById(&quot;3dq3y9&quot;);
+       var containerDiv = document.getElementById(&quot;0m1sRp&quot;);
        
                var toolbar = null;
                var plotContainer = containerDiv;
