+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/workflow/CompareAuditsWithUndervotes.kt
@@ -8,7 +8,7 @@ import kotlin.test.Test
 
 class CompareAuditsWithUndervotes {
     val nruns = 100  // number of times to run workflow
-    val nsimEst = 100  // number of times to run simulation
+    val nsimEst = 10  // number of times to run simulation
     val name = "auditsWithUndervotes"
     val dirName = "/home/stormy/temp/samples/$name"
     val mvrFuzzPct = .01
@@ -17,8 +17,7 @@ class CompareAuditsWithUndervotes {
 
     @Test
     fun genAuditWithUnderVotesPlots() {
-        val cvrPercent = .50
-        val undervotes = listOf(.00, .05, .10, .15, .20, .25, .30, .35, .40, .45, .50)
+        val undervotes = listOf(.00, .10, .20, .30, .40, .50)
         val stopwatch = Stopwatch()
 
         val tasks = mutableListOf<ConcurrentTaskG<List<WorkflowResult>>>()
@@ -34,9 +33,12 @@ class CompareAuditsWithUndervotes {
                 parameters=mapOf("nruns" to nruns, "undervote" to undervote, "cat" to "clca"))
             tasks.add(RepeatedWorkflowRunner(nruns, clcaGenerator))
 
-            val oneauditGenerator = OneAuditSingleRoundAuditTaskGenerator(N, margin, undervote, 0.0, cvrPercent=cvrPercent, mvrsFuzzPct=mvrFuzzPct,
-                auditConfigIn=AuditConfig(AuditType.ONEAUDIT, true, nsimEst = nsimEst),
-                parameters=mapOf("nruns" to nruns, "undervote" to undervote, "cat" to "oneaudit"))
+            val oneauditGenerator = OneAuditSingleRoundAuditTaskGenerator(
+                N, margin, undervote, 0.0, cvrPercent = .99, mvrsFuzzPct=mvrFuzzPct,
+                parameters=mapOf("nruns" to nruns, "undervote" to undervote, "cat" to "oneaudit99"),
+                auditConfigIn = AuditConfig(AuditType.ONEAUDIT, true,
+                    oaConfig = OneAuditConfig(strategy=OneAuditStrategyType.max99))
+            )
             tasks.add(RepeatedWorkflowRunner(nruns, oneauditGenerator))
         }
 
@@ -52,12 +54,10 @@ class CompareAuditsWithUndervotes {
 
     @Test
     fun regenPlots() {
-        val name = "auditsWithUndervotes"
-        val dirName = "/home/stormy/temp/workflow/$name"
-
-        val subtitle = "margin=${margin} Nc=${N} nruns=${nruns}"
+        val subtitle = "margin=${margin} Nc=${N} nruns=${nruns}  mvrFuzz=${mvrFuzzPct}"
         showSampleSizesVsUndervotePct(dirName, name, subtitle, ScaleType.Linear, catName="auditType")
         showSampleSizesVsUndervotePct(dirName, name, subtitle, ScaleType.LogLinear, catName="auditType")
+        showSampleSizesVsUndervotePct(dirName, name, subtitle, ScaleType.LogLog, catName="auditType")
     }
 
     fun showSampleSizesVsUndervotePct(dirName: String, name:String, subtitle: String, scaleType: ScaleType,
